old-tests/submission/W3C/domhtml/HTMLAnchorElement04.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-a-element.html#dom-a-href" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<a href="/support/images/yellow-rectangle.png">link</a>
</p>


<script>
var domtest = async_test("The href attribute contains the URL of the linked resource.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vhref;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vhref = testNode.href;

      assert_uri_equals(document.location.protocol,null,document.location.host,"yellow-rectangle.png",null,null,null,true,vhref);

    domtest.done();
}));

</script>
<p>The href attribute contains the URL of the linked resource. 

    Retrieve the href attribute and examine its value.</p>
<div id="log"></div>
</body></html>
